add ci test
