--- interactive_engine/compiler/src/main/java/com/alibaba/graphscope/common/ir/rel/LoadCSVTableScan.java
@@ -1,84 +1,0 @@
-/*
- *
- *  * Copyright 2020 Alibaba Group Holding Limited.
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  * http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
- *
- */
-
-package com.alibaba.graphscope.common.ir.rel;
-
-import com.alibaba.graphscope.common.ir.meta.schema.LoadCSVTable;
-import com.alibaba.graphscope.common.ir.rel.graph.AbstractBindableTableScan;
-import com.alibaba.graphscope.common.ir.rel.type.TableConfig;
-import com.google.common.collect.ImmutableList;
-
-import org.apache.calcite.plan.GraphOptCluster;
-import org.apache.calcite.plan.RelOptCluster;
-import org.apache.calcite.plan.RelOptTable;
-import org.apache.calcite.prepare.RelOptTableImpl;
-import org.apache.calcite.rel.RelNode;
-import org.apache.calcite.rel.RelShuttle;
-import org.apache.calcite.rel.hint.RelHint;
-import org.apache.calcite.rel.type.RelDataType;
-import org.apache.calcite.rel.type.RelDataTypeFieldImpl;
-import org.apache.calcite.rel.type.RelRecordType;
-import org.checkerframework.checker.nullness.qual.Nullable;
-
-import java.util.List;
-
-/**
- * A table scan operator for loading data from CSV table
- */
-public class LoadCSVTableScan extends AbstractBindableTableScan {
-    private final LoadCSVTable csvTable;
-
-    public LoadCSVTableScan(
-            GraphOptCluster cluster,
-            List<RelHint> hints,
-            LoadCSVTable csvTable,
-            @Nullable String aliasName) {
-        super(cluster, hints, createTableConfig(cluster, csvTable), aliasName);
-        this.csvTable = csvTable;
-    }
-
-    private static TableConfig createTableConfig(RelOptCluster cluster, LoadCSVTable csvTable) {
-        RelOptTable optTable =
-                RelOptTableImpl.create(
-                        null,
-                        csvTable.getRowType(cluster.getTypeFactory()),
-                        csvTable,
-                        csvTable.getName());
-        return new TableConfig(ImmutableList.of(optTable));
-    }
-
-    @Override
-    public RelDataType deriveRowType() {
-        return new RelRecordType(
-                ImmutableList.of(
-                        new RelDataTypeFieldImpl(
-                                getAliasName(), getAliasId(), getTable().getRowType())));
-    }
-
-    @Override
-    public RelNode accept(RelShuttle shuttle) {
-        if (shuttle instanceof GraphShuttle) {
-            return ((GraphShuttle) shuttle).visit(this);
-        }
-        return shuttle.visit(this);
-    }
-
-    public LoadCSVTable getCSVTable() {
-        return this.csvTable;
-    }
-}
