+++ interactive_engine/compiler/src/main/java/com/alibaba/graphscope/common/ir/runtime/write/Request.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2020 Alibaba Group Holding Limited.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.graphscope.common.ir.runtime.write;
+
+public class Request<T> {
+    public final Type type;
+    public final T request;
+
+    public Request(Type type, T request) {
+        this.type = type;
+        this.request = request;
+    }
+
+    @Override
+    public String toString() {
+        return "Request{" + "type=" + type + ", request=" + request + '}';
+    }
+}
